requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-APScheduler==1.13.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
gunicorn==22.0.0
redis==5.2.1
openpyxl==3.1.5
python-dotenv==1.0.1
python-dateutil==2.9.0
Werkzeug==3.1.3
APScheduler==3.10.4
email-validator==2.2.0
